refactor: one type per file + folderized structure; cleaner Mediator internals
