+++ src/ConduitR.Abstractions/Handlers/INotificationHandler.cs
@@ -3,12 +3,9 @@ using System.Threading.Tasks;
 
 namespace ConduitR.Abstractions;
 
-/// <summary>Marker interface for a publish/subscribe notification.</summary>
-public interface INotification {}
-
 /// <summary>Handler for notifications.</summary>
 public interface INotificationHandler<in TNotification>
     where TNotification : INotification
 {
-    Task Handle(TNotification notification, CancellationToken cancellationToken);
+    Task Handle(TNotification notification, System.Threading.CancellationToken cancellationToken);
 }
