+++ src/ConduitR.Abstractions/Handlers/IRequestHandler.cs
