+++ src/ConduitR.Abstractions/Notifications/INotification.cs
@@ -1,0 +1,4 @@
+namespace ConduitR.Abstractions;
+
+/// <summary>Marker interface for a publish/subscribe notification.</summary>
+public interface INotification {}
